app/requirements.txt

# Python dependencies
google-adk
fastapi
uvicorn
pyyaml
sqlalchemy 